refactor: relocate economy preview test support under test/support (#3876)
Moves economy_stockpile_preview_test_support.dart into test/support/ so
the colonizethis_turn test root retains only the three economy preview
test entrypoints required by the wave-2 consolidation AC.
