+++ packages/colonizethis_turn/test/economy_stockpile_preview_recruit_worker_test.dart
@@ -3,7 +3,7 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import 'economy_stockpile_preview_test_support.dart';
+import 'support/economy_stockpile_preview_test_support.dart';
 import 'package:colonizethis_test/game_test_fixtures.dart';
 
 /// Pending [RecruitWorkerOrder] costs feed the production-panel economy
