+++ packages/colonizethis_turn/test/economy_stockpile_preview_test.dart
@@ -2,7 +2,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
-import 'economy_stockpile_preview_test_support.dart';
+import 'support/economy_stockpile_preview_test_support.dart';
 import 'support/economy_stockpile_preview_pending_work_scenarios.dart';
 import 'package:colonizethis_test/game_test_fixtures.dart';
